Nuked old 3d rover
replaced the old 3d rover files with a single component. Also added two libraries: react-three/fiber and react-three/drei
