+++ package.json
@@ -181,6 +181,7 @@
     "@types/react-router-dom": "^5.3.2",
     "@types/react-test-renderer": "^16.9.5",
     "@types/react-vis": "^1.11.11",
+    "@types/three": "^0.162.0",
     "@types/webpack-env": "^1.16.3",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
@@ -245,6 +246,8 @@
     "yarn-deduplicate": "^3.1.0"
   },
   "dependencies": {
+    "@react-three/drei": "7.5.1",
+    "@react-three/fiber": "7.0.6",
     "ansi-regex": "^6.0.1",
     "double-ended-queue": "^2.1.0-0",
     "electron-db": "^0.15.7",
@@ -269,7 +272,8 @@
     "react-vis": "^1.11.7",
     "regenerator-runtime": "^0.13.9",
     "source-map-support": "^0.5.20",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.3",
+    "three": "0.130.0"
   },
   "devEngines": {
     "node": ">=10.x",
