+++ src/RID/RID.tsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import CSS from 'csstype';
 import Cameras from '../Core/components/Cameras';
 import Map from '../RED/components/Map';
-import ThreeDRover from '../Core/components/ThreeDRover';
+import Accelerometer from '../Core/components/Accelerometer';
 import Angular from '../RAM/Arm/components/Angular';
 import ControlFeatures from '../RAM/Arm/components/ControlFeatures';
 import IK from '../RAM/Arm/components/IK';
@@ -173,7 +173,7 @@ class RoverImageryDisplay extends Component<IProps, IState> {
       display: (
         <div style={{ flexGrow: 1, height: '100%' }}>
           {this.buttons}
-          <ThreeDRover style={submod} zoom={30} />
+          <Accelerometer style={submod} zoom={30} />
         </div>
       ),
       displayed: '3D Rover',
